firmware: Send 'I' hist with interrupts enabled
For *intermediate* results, we can live with the glitches
that may be caused by ISR(ADC_vect) updating a counter
value while send_histogram() is reading it.

For the final results (*abort*ed or *done* type histograms),
interrupts will always be disabled and the 'A' and 'D' results
thus will always be glitch free.

Keeping interrupts enabled also avoids the issue of needing
to properly pause the timer while send_histogram() is taking
its sweet time with the measurement disabled.