fw,host: Always send "state" packets
In almost all firmware FSM states, send state packets
on entering the state. The state packets contain exactly
the name of the state as in the specs.

The specs now specify the capital letter states which
are sent and the lower case letter states which are not.

Caution, this kind of breaks the emulator code which
has not been adapted yet.