firmware: Clean FSM implementation using state type
The source code looks cleaner and is easier to read,
compared to the direct approach of nesting and chaining
a few while loop and switch statements and uart_getc()
calls.

However, it adds about 20 source lines of code and
about 130 bytes of binary code size!

The binary code increase and resulting firmware image
transmission times of >7% is probably not worth it.