firmware: Set F_CPU via build system (16 or 18.432MHz)
We need to support two F_CPU values, and putting that difference
into settings.mk instead of the source code makes the code easier
to maintain.

To use this, add a line to firmware/settings.mk like
    F_CPU = 16000000UL
or
    F_CPU = 18432000UL

The default F_CPU value (if none given in settings.mk)
is 16000000UL, i.e. 16MHz.