firmware build: Do string replacement in ELF binary
Do string replacement in ELF binary after linking to avoid
useless runtime string formatting.  This is an ugly mechanism,
but why do string formatting on the embedded device at runtime
when we can determine the string at build time?

Has the time series firmware print the table size in elements
and the element size in bit.