firmware: Move static strings to program space
Move all static strings to pgmspace, i.e. to flash memory.

Yes, this adds a total of 100 bytes to the firmware image
that need to be flashed.

However, this also saves us 172 bytes in the SRAM for the
geiger-ts personality - and those 172 bytes we can put to
good use in the data table.