proto: Put msmt params in front, then token data
Transmit measurement params in the front of the parameter
block, and then append an arbitrary number of bytes to be
stored in fw as token (max number is 16).

Firmware will send back the additional bytes unchanged.

To distinguish the sent and received data on the screen,
prepend the lines with ">" and "<" markers, respectively.