fw: Globally enable interrupts on fw startup
Set up the events causing ISRs only on measurement start,
not on firmware startup. This makes it possible to globally
enable interrupts on startup (which we need in preparation
for future code).

To keep the changes simple, we we just move the hardware
initialization from firmware boot (.initN section) to
measurement start (personality_start_measurement_sram()
function).