fw: inline trivial checksum functions
This reduces code size as the function bodies are smaller than
the code to call them.

Also this makes checksum_matches() use less code by itself.

Apparently, this was an 16bit(!) comparison of a and b:

   uint8_t a, b;
   return (a == b);

The new code reduces this to an 8bit comparison:

   uint8_t a, b;
   const uint8 d = a-b;
   return (0 == d);

Note that we are NOT moving the receive checksum into a local variable
in the main() loop - that runs the very real risk of the avr-gcc
optimizations messing up the registers again.