Frame format FMPX: consistent TLV field, CRC16
Now frames sent in both directions all start with
the four byte magic header and a single letter
designating the frame type.

Lower case letters are commands sent to the device,
upper case letters are data frames sent to the host.

Only after the command/type byte, the length field
starts (one byte for the commands, and two bytes for
the data frames).