fw: report size of data_table for each personality
Report the size and composition of data_table (number of bytes,
number of elements and element size in bits) for each FW personality.

This wastes between 8 and 32 bytes of PROGMEM flash memory in both
the firmware image and on the MCU for the table_element_bits array.
The array size in bytes is used to communicated the element size in
bits.

There are no runtime costs associated with this method.