fw: data_table memory layout adapted to upstream memory layout
Upstream statically allocates all data before _end.

So we insert our data_table at _end and move _end to data_table_end.

As the __heap_start is at _end, we move that as well.

Upstream uses no __heap_end (heap and stack are growing against each
other dynamically), so we get rid of __heap_end as well to avoid
confusion.

Oh, and we do not even use any of the heap, so we have the heap are
between __heap_start and __heap_end as an additional safety margin
against the stack dynamically growing too much.

And we check that we do not step over that safety margin.